+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2022-11
 
+- create status updates upon mission submission
 - ping request now with MxmInfo payload (`POST /mxmInfo` removed in Provider API)
 
 2022-10